requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
pandas
pathlib
prophet
resend
slowapi
pydantic[email]
firebase-admin
httpx
sendgrid
reportlab
azure-ai-contentsafety